Move logical lattice operations to theri own file
